feat: moved code inside collective domain
