+++ src/renderer/domains/collectives/_lib/types.ts
