+++ src/renderer/domains/collectives/init.ts
@@ -1,16 +1,16 @@
 import { combine } from 'effector';
 
-import { combineStores } from './lib/helpers';
-import { membersDomainModel } from './model/members/model';
-import { membersService } from './model/members/service';
-import { referendumDomainModel } from './model/referendum/model';
-import { referendumService } from './model/referendum/service';
-import { referendumMetaModel } from './model/referendumMeta/model';
-import { tracksDomainModel } from './model/tracks/model';
-import { tracksService } from './model/tracks/service';
-import { votesDomainModel } from './model/votes/model';
-import { votingDomainModel } from './model/voting/model';
-import { votingService } from './model/voting/service';
+import { combineStores } from './_lib/helpers';
+import { membersDomainModel } from './members/model';
+import { membersService } from './members/service';
+import { referendumDomainModel } from './referendum/model';
+import { referendumService } from './referendum/service';
+import { referendumMetaModel } from './referendumMeta/model';
+import { tracksDomainModel } from './tracks/model';
+import { tracksService } from './tracks/service';
+import { votesDomainModel } from './votes/model';
+import { votingDomainModel } from './voting/model';
+import { votingService } from './voting/service';
 
 const $store = combine(
   {
