+++ src/renderer/domains/collectives/members/types.ts
