+++ src/renderer/domains/collectives/referendum/mapper.ts
