+++ src/renderer/domains/collectives/referendum/types.ts
