+++ src/renderer/domains/collectives/referendumMeta/types.ts
