+++ src/renderer/domains/collectives/votes/mapper.ts
