+++ src/renderer/domains/collectives/votes/types.ts
