+++ src/renderer/domains/collectives/voting/types.ts
@@ -1,7 +1,7 @@
 import { type Transaction, type XOR } from '@/shared/core';
 import { type ReferendumId } from '@/shared/pallet/referenda';
 import { type AccountId } from '@/shared/polkadotjs-schemas';
-import { type CollectivePalletsType } from '../../lib/types';
+import { type CollectivePalletsType } from '../_lib/types';
 
 export type Voting = XOR<{ aye: number }, { nay: number }> & {
   accountId: AccountId;
